Refactor: Organize project structure with common and utils modules
