+++ modules/game_time_calculator.py
@@ -1,9 +1,9 @@
 import tkinter as tk
 from tkinter import ttk
-from modules.game_time_tab_poomsae import PoomsaeTab
-from modules.game_time_tab_kyorugi import KyorugiTab
-from version import __version__ as app_version
-from version import __build_date__ as app_date
+from .game_time_tab_poomsae import PoomsaeTab
+from .game_time_tab_kyorugi import KyorugiTab
+from ..common.version import __version__ as app_version
+from ..common.version import __build_date__ as app_date
 
 class GameTimeCalculator(tk.Toplevel):
     def __init__(self, master):
